+++ client/components/WBmodules/Rectangle.js
