+++ client/components/index.js
@@ -19,7 +19,7 @@ export {default as Students} from './Students'
 export {default as Profile} from './Profile'
 export {default as Lessons} from './Lessons'
 export {default as ManageAssignments} from './ManageAssignments'
-export {default as Canvas} from './Canvas'
+export {default as WhiteBoard} from './WhiteBoard'
 export {default as Grades} from './Grades'
 export {default as Verification} from './Verification'
 export {default as OneStudentGrades} from './OneStudentGrades'
